enable compiler warnings & fix them

+++ main.c
@@ -7,7 +7,7 @@ char* user_input;
 
 int dump_tokens(void) {
     while (token->kind != TK_EOF) {
-        printf("%.*s\t%d\n", token->len, token->str, token->kind);
+        printf("%.*s\t%d\n", (int)(token->len), token->str, token->kind);
         token = token->next;
     }
     return 0;
